Refactoring customer and supplier invoice
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C53776

+++ TOneV2/Code/TOneV2/TOne.WhS.Invoice.Entities/CustomerInvoiceDetails.cs
@@ -17,7 +17,7 @@ namespace TOne.WhS.Invoice.Entities
         public string OriginalSaleCurrency { get; set; }
         public int CountryId { get; set; }
         public int SupplierId { get; set; }
-        public int SupplierZoneId { get; set; }
+        public long SupplierZoneId { get; set; }
         public string CustomerCurrency { get; set; }
 
         public CustomerInvoiceDetails() { }
